User specific content with JWT claims

+++ Controllers/V1/PostsController.cs
@@ -10,6 +10,7 @@ using PostChan.Contracts.V1.Responses;
 using PostChan.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using PostChan.Extensions;
 
 namespace PostChan.Controllers.V1
 {
@@ -43,7 +44,11 @@ namespace PostChan.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post 
+            {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId(),
+            };
 
             await _postService.CreatePostAsync(post);
 
@@ -58,6 +63,13 @@ namespace PostChan.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute]Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own that post" });
+            }
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
@@ -69,11 +81,15 @@ namespace PostChan.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromBody] UpdatePostRequest request, [FromRoute] Guid postId)
         {
-            var post = new Post
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
             {
-                Id = postId,
-                Name = request.Name
-            };
+                return BadRequest(new { error = "You do not own that post" });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
             
             if (await _postService.UpdatePostAsync(post))
                 return Ok(post);
